+++ config/config.exs
@@ -1,3 +1,7 @@
-use Mix.Config
-# config :blake3, simd_mode: "std"
-# config :blake3, rayon: :false
+import Config
+
+config :blake3, Blake3.Native,
+  mode: :release,
+  features: MixBlake3.Project.config_features()
+
+import_config("#{Mix.env()}.exs")