+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## v0.6.0
+
+- Enhancements
+
+  - updated blake3 version to 1.3.3
+  - updated rustler to 0.22.2
+
+- Other
+  - add github actions release workflow
+  - fix CI
+  - Release version injection
+
+## v0.5.0 (unreleased)
+
+- Enhancements
+
+  - updated blake3 version to 1.0.0
+  - updated rustler to 0.22.1
+  - refactored Rust crate to use the rustler::nif macro
+
+- Breaking changes
+
+  - `update_with_join` has been replacted with `update_rayon` to reflect the
+    same change in the upstream `blake3` crate. This function is only available
+    when the `rayon` feature flag is enabled, and will panic otherwise.
+
 ## v0.4.0
 
 - Enhancements